refactor: 오류 해결
