+++ helper-persistence/src/main/java/org/example/persistence/gradeInfo/repository/GradeInfoJpaRepository.java
@@ -1,14 +1,14 @@
 package org.example.persistence.gradeInfo.repository;
 
-import org.example.persistence.gradeInfo.entity.GradeInfo;
-import org.example.persistence.gradeInfo.entity.GradeInfoId;
+import org.example.persistence.student.entity.GradeInfoId;
+import org.example.persistence.student.entity.GradeInfoJpaEntity;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface GradeInfoJpaRepository extends CrudRepository<GradeInfo, GradeInfoId> {
+public interface GradeInfoJpaRepository extends CrudRepository<GradeInfoJpaEntity, GradeInfoId> {
 
-    Optional<GradeInfo> findByClassroomAndGrade(String classroom, String grade);
+    Optional<GradeInfoJpaEntity> findByClassroomAndGrade(String classroom, String grade);
 
     Boolean existsByGradeAndClassroom(String grade, String classroom);
 }
